Refactor p2p_sync
