+++ src/vbk/p2p_sync.hpp
@@ -13,8 +13,7 @@
 #include <node/context.h>
 #include <rpc/blockchain.h>
 #include <vbk/pop_common.hpp>
-
-#include "veriblock/mempool.hpp"
+#include <veriblock/mempool.hpp>
 
 namespace VeriBlock {
 
@@ -37,36 +36,9 @@ struct PopDataNodeState {
     std::map<typename T::id_t, PopP2PState>& getMap();
 };
 
-template <>
-inline std::map<altintegration::ATV::id_t, PopP2PState>& PopDataNodeState::getMap<altintegration::ATV>()
-{
-    return atv_state;
-}
+PopDataNodeState& getPopDataNodeState(const NodeId& id);
 
-template <>
-inline std::map<altintegration::VTB::id_t, PopP2PState>& PopDataNodeState::getMap<altintegration::VTB>()
-{
-    return vtb_state;
-}
-
-template <>
-inline std::map<altintegration::VbkBlock::id_t, PopP2PState>& PopDataNodeState::getMap<altintegration::VbkBlock>()
-{
-    return vbk_blocks_state;
-}
-
-/** Map maintaining peer PopData state  */
-extern std::map<NodeId, std::shared_ptr<PopDataNodeState>> mapPopDataNodeState GUARDED_BY(cs_main);
-
-inline PopDataNodeState& getPopDataNodeState(const NodeId& id) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    std::shared_ptr<PopDataNodeState> val = mapPopDataNodeState[id];
-    if (val == nullptr) {
-        val = std::make_shared<PopDataNodeState>();
-        mapPopDataNodeState[id] = val;
-    }
-    return *val;
-}
+void erasePopDataNodeState(const NodeId& id);
 
 } // namespace p2p
